Исправлено добавление расстояний в матрицу весов. Добавление перемещение зомби, но пока работает криво - идут в одну точку (10,10) примерно. Время срабатывания таймера изменено 50мс-> 2000мс

+++ Human.cs
@@ -8,10 +8,10 @@ namespace CodeVsZombies2
 {
     class Human
     {
-        public int x;
-        public int y;
+        public double x;
+        public double y;
 
-        public Human(int x, int y)
+        public Human(double x, double y)
         {
             this.x = x;
             this.y = y;
